Rustfmt, move SpinTimer, make dotstar peripheral take a timer, update .travis.yml
