+++ .travis.yml
@@ -10,7 +10,7 @@ env:
   - CRATE=boards/metro_m0 EXAMPLES="--example=blinky_basic --example=blinky_rtfm"
   - CRATE=boards/metro_m4 FEATURES="--features=unproven"
   - CRATE=boards/feather_m0 FEATURES="--features=unproven"
-  - CRATE=boards/feather_m4
+  - CRATE=boards/feather_m4 FEATURES="--features=unproven"
   - CRATE=boards/gemma_m0
   - CRATE=boards/itsybitsy_m0
   - CRATE=boards/itsybitsy_m4
