+++ boards/itsybitsy_m4/examples/dotstar.rs
@@ -9,6 +9,7 @@ use hal::delay::Delay;
 use hal::entry;
 use hal::pac::{CorePeripherals, Peripherals};
 use hal::prelude::*;
+use hal::timer::SpinTimer;
 
 use smart_leds::RGB8;
 use smart_leds_trait::SmartLedsWrite;
@@ -28,7 +29,7 @@ fn main() -> ! {
 
     let mut pins = hal::Pins::new(peripherals.PORT).split();
 
-    let mut rgb = hal::dotstar_bitbang(pins.dotstar, &mut pins.port);
+    let mut rgb = hal::dotstar_bitbang(pins.dotstar, &mut pins.port, SpinTimer::new(12));
     let off: [RGB8; 1] = [RGB8 { r: 0, g: 0, b: 0 }];
     let red: [RGB8; 1] = [RGB8 { r: 100, g: 0, b: 0 }];
     let green: [RGB8; 1] = [RGB8 { r: 0, g: 100, b: 0 }];
@@ -38,8 +39,8 @@ fn main() -> ! {
 
     loop {
         rgb.write(red.iter().cloned()).unwrap();
-        delay.delay_ms(200u8);
+        delay.delay_ms(60u8);
         rgb.write(green.iter().cloned()).unwrap();
-        delay.delay_ms(200u8);
+        delay.delay_ms(60u8);
     }
 }
